Move test_datetime.py
