+++ python/cudf/cudf/tests/series/test_binops.py
@@ -501,3 +501,290 @@ def test_concatenate_list_with_nonlist():
     gser2 = cudf.Series(["a", "b", "c"])
     with pytest.raises(TypeError):
         gser1 + gser2
+
+
+def test_datetime_series_binops_pandas(
+    datetime_types_as_str, datetime_types_as_str2
+):
+    dti = pd.date_range("20010101", "20020215", freq="400h", name="times")
+    pd_data_1 = pd.Series(dti)
+    pd_data_2 = pd_data_1
+    gdf_data_1 = cudf.Series(pd_data_1).astype(datetime_types_as_str)
+    gdf_data_2 = cudf.Series(pd_data_2).astype(datetime_types_as_str2)
+    assert_eq(pd_data_1, gdf_data_1.astype("datetime64[ns]"))
+    assert_eq(pd_data_2, gdf_data_2.astype("datetime64[ns]"))
+    assert_eq(pd_data_1 < pd_data_2, gdf_data_1 < gdf_data_2)
+    assert_eq(pd_data_1 > pd_data_2, gdf_data_1 > gdf_data_2)
+    assert_eq(pd_data_1 == pd_data_2, gdf_data_1 == gdf_data_2)
+    assert_eq(pd_data_1 <= pd_data_2, gdf_data_1 <= gdf_data_2)
+    assert_eq(pd_data_1 >= pd_data_2, gdf_data_1 >= gdf_data_2)
+
+
+def test_datetime_series_binops_numpy(
+    datetime_types_as_str, datetime_types_as_str2
+):
+    dti = pd.date_range("20010101", "20020215", freq="400h", name="times")
+    pd_data_1 = pd.Series(dti)
+    pd_data_2 = pd_data_1
+    gdf_data_1 = cudf.Series(pd_data_1).astype(datetime_types_as_str)
+    gdf_data_2 = cudf.Series(pd_data_2).astype(datetime_types_as_str2)
+    np_data_1 = np.array(pd_data_1).astype(datetime_types_as_str)
+    np_data_2 = np.array(pd_data_2).astype(datetime_types_as_str2)
+    np.testing.assert_equal(np_data_1, gdf_data_1.to_numpy())
+    np.testing.assert_equal(np_data_2, gdf_data_2.to_numpy())
+    np.testing.assert_equal(
+        np.less(np_data_1, np_data_2), (gdf_data_1 < gdf_data_2).to_numpy()
+    )
+    np.testing.assert_equal(
+        np.greater(np_data_1, np_data_2), (gdf_data_1 > gdf_data_2).to_numpy()
+    )
+    np.testing.assert_equal(
+        np.equal(np_data_1, np_data_2), (gdf_data_1 == gdf_data_2).to_numpy()
+    )
+    np.testing.assert_equal(
+        np.less_equal(np_data_1, np_data_2),
+        (gdf_data_1 <= gdf_data_2).to_numpy(),
+    )
+    np.testing.assert_equal(
+        np.greater_equal(np_data_1, np_data_2),
+        (gdf_data_1 >= gdf_data_2).to_numpy(),
+    )
+
+
+@pytest.mark.parametrize(
+    "data",
+    [
+        pd.date_range("20010101", "20020215", freq="400h", name="times"),
+        pd.date_range(
+            "20010101", freq="243434324423423234ns", name="times", periods=10
+        ),
+    ],
+)
+def test_dt_ops(data):
+    pd_data = pd.Series(data)
+    gdf_data = cudf.Series(data)
+
+    assert_eq(pd_data == pd_data, gdf_data == gdf_data)
+    assert_eq(pd_data < pd_data, gdf_data < gdf_data)
+    assert_eq(pd_data > pd_data, gdf_data > gdf_data)
+
+
+@pytest.mark.parametrize(
+    "data",
+    [
+        [1, 2, 3, 4, 10, 100, 20000],
+        [None] * 7,
+        [10, 20, 30, None, 100, 200, None],
+        [3223.234, 342.2332, 23423.23, 3343.23324, 23432.2323, 242.23, 233],
+    ],
+)
+@pytest.mark.parametrize(
+    "other",
+    [
+        [1, 2, 3, 4, 10, 100, 20000],
+        [None] * 7,
+        [10, 20, 30, None, 100, 200, None],
+        [3223.234, 342.2332, 23423.23, 3343.23324, 23432.2323, 242.23, 233],
+        datetime.datetime(1993, 6, 22, 13, 30),
+        datetime.datetime(2005, 1, 22, 10, 00),
+        np.datetime64("2005-02"),
+        np.datetime64("2005-02-25"),
+        np.datetime64("2005-02-25T03:30"),
+        np.datetime64("nat"),
+        # TODO: https://github.com/pandas-dev/pandas/issues/52295
+    ],
+)
+def test_datetime_subtract(
+    data, other, datetime_types_as_str, datetime_types_as_str2
+):
+    gsr = cudf.Series(data, dtype=datetime_types_as_str)
+    psr = gsr.to_pandas()
+
+    if isinstance(other, np.datetime64):
+        gsr_other = other
+        psr_other = other
+    else:
+        gsr_other = cudf.Series(other, dtype=datetime_types_as_str2)
+        psr_other = gsr_other.to_pandas()
+
+    expected = psr - psr_other
+    actual = gsr - gsr_other
+
+    assert_eq(expected, actual)
+
+    expected = psr_other - psr
+    actual = gsr_other - gsr
+
+    assert_eq(expected, actual)
+
+
+@pytest.mark.parametrize(
+    "data",
+    [
+        [1000000, 200000, 3000000],
+        [1000000, 200000, None],
+        [],
+        [None],
+        [None, None, None, None, None],
+        [12, 12, 22, 343, 4353534, 435342],
+        np.array([10, 20, 30, None, 100]),
+        cp.asarray([10, 20, 30, 100]),
+        [1000000, 200000, 3000000],
+        [1000000, 200000, None],
+        [1],
+        [12, 11, 232, 223432411, 2343241, 234324, 23234],
+        [12, 11, 2.32, 2234.32411, 2343.241, 23432.4, 23234],
+        [1.321, 1132.324, 23223231.11, 233.41, 0.2434, 332, 323],
+        [12, 11, 2.32, 2234.32411, 2343.241, 23432.4, 23234],
+    ],
+)
+@pytest.mark.parametrize(
+    "other_scalars",
+    [
+        datetime.timedelta(days=768),
+        datetime.timedelta(seconds=768),
+        datetime.timedelta(microseconds=7),
+        datetime.timedelta(minutes=447),
+        datetime.timedelta(hours=447),
+        datetime.timedelta(weeks=734),
+        np.timedelta64(4, "s"),
+        np.timedelta64(456, "D"),
+        np.timedelta64(46, "h"),
+        np.timedelta64("nat"),
+        np.timedelta64(1, "s"),
+        np.timedelta64(1, "ms"),
+        np.timedelta64(1, "us"),
+        np.timedelta64(1, "ns"),
+    ],
+)
+@pytest.mark.parametrize("op", ["add", "sub"])
+def test_datetime_series_ops_with_scalars(
+    data, other_scalars, datetime_types_as_str, op
+):
+    gsr = cudf.Series(data=data, dtype=datetime_types_as_str)
+    psr = gsr.to_pandas()
+
+    if op == "add":
+        expected = psr + other_scalars
+        actual = gsr + other_scalars
+    elif op == "sub":
+        expected = psr - other_scalars
+        actual = gsr - other_scalars
+
+    assert_eq(expected, actual)
+
+    if op == "add":
+        expected = other_scalars + psr
+        actual = other_scalars + gsr
+
+        assert_eq(expected, actual)
+
+    elif op == "sub":
+        assert_exceptions_equal(
+            lfunc=operator.sub,
+            rfunc=operator.sub,
+            lfunc_args_and_kwargs=([other_scalars, psr],),
+            rfunc_args_and_kwargs=([other_scalars, gsr],),
+        )
+
+
+@pytest.mark.parametrize("data", ["20110101", "20120101", "20130101"])
+@pytest.mark.parametrize("other_scalars", ["20110101", "20120101", "20130101"])
+def test_datetime_series_cmpops_with_scalars(
+    data, other_scalars, datetime_types_as_str, comparison_op
+):
+    gsr = cudf.Series(data=data, dtype=datetime_types_as_str)
+    psr = gsr.to_pandas()
+
+    expect = comparison_op(psr, other_scalars)
+    got = comparison_op(gsr, other_scalars)
+
+    assert_eq(expect, got)
+
+
+def test_datetime_invalid_ops():
+    sr = cudf.Series([1, 2, 3], dtype="datetime64[ns]")
+    psr = sr.to_pandas()
+
+    assert_exceptions_equal(
+        lfunc=operator.add,
+        rfunc=operator.add,
+        lfunc_args_and_kwargs=([psr, pd.Timestamp(1513393355.5, unit="s")],),
+        rfunc_args_and_kwargs=([sr, pd.Timestamp(1513393355.5, unit="s")],),
+    )
+
+    assert_exceptions_equal(
+        lfunc=operator.truediv,
+        rfunc=operator.truediv,
+        lfunc_args_and_kwargs=([psr, pd.Timestamp(1513393355.5, unit="s")],),
+        rfunc_args_and_kwargs=([sr, pd.Timestamp(1513393355.5, unit="s")],),
+    )
+
+    assert_exceptions_equal(
+        lfunc=operator.add,
+        rfunc=operator.add,
+        lfunc_args_and_kwargs=([psr, psr],),
+        rfunc_args_and_kwargs=([sr, sr],),
+    )
+
+    assert_exceptions_equal(
+        lfunc=operator.floordiv,
+        rfunc=operator.floordiv,
+        lfunc_args_and_kwargs=([psr, psr],),
+        rfunc_args_and_kwargs=([sr, sr],),
+    )
+
+    assert_exceptions_equal(
+        lfunc=operator.floordiv,
+        rfunc=operator.floordiv,
+        lfunc_args_and_kwargs=([psr, pd.Timestamp(1513393355.5, unit="s")],),
+        rfunc_args_and_kwargs=([sr, pd.Timestamp(1513393355.5, unit="s")],),
+    )
+
+    assert_exceptions_equal(
+        lfunc=operator.add,
+        rfunc=operator.add,
+        lfunc_args_and_kwargs=([psr, 1],),
+        rfunc_args_and_kwargs=([sr, 1],),
+    )
+
+    assert_exceptions_equal(
+        lfunc=operator.truediv,
+        rfunc=operator.truediv,
+        lfunc_args_and_kwargs=([psr, "a"],),
+        rfunc_args_and_kwargs=([sr, "a"],),
+    )
+
+    assert_exceptions_equal(
+        lfunc=operator.mul,
+        rfunc=operator.mul,
+        lfunc_args_and_kwargs=([psr, 1],),
+        rfunc_args_and_kwargs=([sr, 1],),
+    )
+
+
+def test_datetime_binop_tz_timestamp(comparison_op):
+    s = cudf.Series([1, 2, 3], dtype="datetime64[ns]")
+    pd_tz_timestamp = pd.Timestamp("1970-01-01 00:00:00.000000001", tz="utc")
+    with pytest.raises(NotImplementedError):
+        comparison_op(s, pd_tz_timestamp)
+
+    date_scalar = datetime.datetime.now(datetime.timezone.utc)
+    with pytest.raises(NotImplementedError):
+        comparison_op(s, date_scalar)
+
+
+def test_datetime_series_cmpops_pandas_compatibility(comparison_op):
+    data1 = ["20110101", "20120101", None, "20140101", None]
+    data2 = ["20110101", "20120101", "20130101", None, None]
+    gsr1 = cudf.Series(data=data1, dtype="datetime64[ns]")
+    psr1 = gsr1.to_pandas()
+
+    gsr2 = cudf.Series(data=data2, dtype="datetime64[ns]")
+    psr2 = gsr2.to_pandas()
+
+    expect = comparison_op(psr1, psr2)
+    with cudf.option_context("mode.pandas_compatible", True):
+        got = comparison_op(gsr1, gsr2)
+
+    assert_eq(expect, got)
