+++ python/cudf/cudf/tests/series/test_np_ufuncs.py
@@ -1,5 +1,5 @@
 # Copyright (c) 2025, NVIDIA CORPORATION.
-
+import datetime
 import operator
 from functools import reduce
 
@@ -208,3 +208,49 @@ def test_ufunc_cudf_series_error_with_out_kwarg():
     cudf_s3 = cudf.Series(data=[0, 0, 0, 0])
     with pytest.raises(TypeError):
         np.add(x1=cudf_s1, x2=cudf_s2, out=cudf_s3)
+
+
+@pytest.mark.parametrize(
+    "data",
+    [
+        [1000000, 200000, 3000000],
+        [1000000, 200000, None],
+        [],
+        [None],
+        [None, None, None, None, None],
+        [12, 12, 22, 343, 4353534, 435342],
+        np.array([10, 20, 30, None, 100]),
+        cp.asarray([10, 20, 30, 100]),
+        [1000000, 200000, 3000000],
+        [1000000, 200000, None],
+        [1],
+        [12, 11, 232, 223432411, 2343241, 234324, 23234],
+        [12, 11, 2.32, 2234.32411, 2343.241, 23432.4, 23234],
+        [1.321, 1132.324, 23223231.11, 233.41, 0.2434, 332, 323],
+        [12, 11, 2.32, 2234.32411, 2343.241, 23432.4, 23234],
+    ],
+)
+@pytest.mark.parametrize(
+    "scalar",
+    [
+        datetime.timedelta(days=768),
+        datetime.timedelta(seconds=768),
+        datetime.timedelta(microseconds=7),
+        np.timedelta64("nat"),
+        np.timedelta64(1, "s"),
+        np.timedelta64(1, "ms"),
+        np.timedelta64(1, "us"),
+        np.timedelta64(1, "ns"),
+    ],
+)
+@pytest.mark.parametrize("op", [np.add, np.subtract])
+def test_datetime_series_ops_with_scalars_misc(
+    data, scalar, datetime_types_as_str, op
+):
+    gsr = cudf.Series(data=data, dtype=datetime_types_as_str)
+    psr = gsr.to_pandas()
+
+    expect = op(psr, scalar)
+    got = op(gsr, scalar)
+
+    assert_eq(expect, got)
