--- python/cudf/cudf/tests/test_datetime.py
@@ -1,422 +1,0 @@
-# Copyright (c) 2019-2025, NVIDIA CORPORATION.
-
-import datetime
-import operator
-
-import cupy as cp
-import numpy as np
-import pandas as pd
-import pytest
-
-import cudf
-from cudf import Series
-from cudf.core._compat import (
-    PANDAS_CURRENT_SUPPORTED_VERSION,
-    PANDAS_VERSION,
-)
-from cudf.testing import assert_eq
-from cudf.testing._utils import (
-    DATETIME_TYPES,
-    assert_exceptions_equal,
-)
-
-
-@pytest.fixture(
-    params=[
-        operator.lt,
-        operator.gt,
-        operator.le,
-        operator.ge,
-        operator.eq,
-        operator.ne,
-    ]
-)
-def op(request):
-    return request.param
-
-
-@pytest.fixture(
-    params=[
-        pd.date_range("20010101", "20020215", freq="400h", name="times"),
-        pd.date_range(
-            "20010101", freq="243434324423423234ns", name="times", periods=10
-        ),
-    ]
-)
-def data(request):
-    return request.param
-
-
-@pytest.mark.parametrize(
-    "lhs_dtype",
-    ["datetime64[s]", "datetime64[ms]", "datetime64[us]", "datetime64[ns]"],
-)
-@pytest.mark.parametrize(
-    "rhs_dtype",
-    ["datetime64[s]", "datetime64[ms]", "datetime64[us]", "datetime64[ns]"],
-)
-def test_datetime_series_binops_pandas(lhs_dtype, rhs_dtype):
-    pd_data_1 = pd.Series(
-        pd.date_range("20010101", "20020215", freq="400h", name="times")
-    )
-    pd_data_2 = pd.Series(
-        pd.date_range("20010101", "20020215", freq="401h", name="times")
-    )
-    gdf_data_1 = Series(pd_data_1).astype(lhs_dtype)
-    gdf_data_2 = Series(pd_data_2).astype(rhs_dtype)
-    assert_eq(pd_data_1, gdf_data_1.astype("datetime64[ns]"))
-    assert_eq(pd_data_2, gdf_data_2.astype("datetime64[ns]"))
-    assert_eq(pd_data_1 < pd_data_2, gdf_data_1 < gdf_data_2)
-    assert_eq(pd_data_1 > pd_data_2, gdf_data_1 > gdf_data_2)
-    assert_eq(pd_data_1 == pd_data_2, gdf_data_1 == gdf_data_2)
-    assert_eq(pd_data_1 <= pd_data_2, gdf_data_1 <= gdf_data_2)
-    assert_eq(pd_data_1 >= pd_data_2, gdf_data_1 >= gdf_data_2)
-
-
-@pytest.mark.parametrize(
-    "lhs_dtype",
-    ["datetime64[s]", "datetime64[ms]", "datetime64[us]", "datetime64[ns]"],
-)
-@pytest.mark.parametrize(
-    "rhs_dtype",
-    ["datetime64[s]", "datetime64[ms]", "datetime64[us]", "datetime64[ns]"],
-)
-def test_datetime_series_binops_numpy(lhs_dtype, rhs_dtype):
-    pd_data_1 = pd.Series(
-        pd.date_range("20010101", "20020215", freq="400h", name="times")
-    )
-    pd_data_2 = pd.Series(
-        pd.date_range("20010101", "20020215", freq="401h", name="times")
-    )
-    gdf_data_1 = Series(pd_data_1).astype(lhs_dtype)
-    gdf_data_2 = Series(pd_data_2).astype(rhs_dtype)
-    np_data_1 = np.array(pd_data_1).astype(lhs_dtype)
-    np_data_2 = np.array(pd_data_2).astype(rhs_dtype)
-    np.testing.assert_equal(np_data_1, gdf_data_1.to_numpy())
-    np.testing.assert_equal(np_data_2, gdf_data_2.to_numpy())
-    np.testing.assert_equal(
-        np.less(np_data_1, np_data_2), (gdf_data_1 < gdf_data_2).to_numpy()
-    )
-    np.testing.assert_equal(
-        np.greater(np_data_1, np_data_2), (gdf_data_1 > gdf_data_2).to_numpy()
-    )
-    np.testing.assert_equal(
-        np.equal(np_data_1, np_data_2), (gdf_data_1 == gdf_data_2).to_numpy()
-    )
-    np.testing.assert_equal(
-        np.less_equal(np_data_1, np_data_2),
-        (gdf_data_1 <= gdf_data_2).to_numpy(),
-    )
-    np.testing.assert_equal(
-        np.greater_equal(np_data_1, np_data_2),
-        (gdf_data_1 >= gdf_data_2).to_numpy(),
-    )
-
-
-def test_dt_ops(data):
-    pd_data = pd.Series(data)
-    gdf_data = Series(data)
-
-    assert_eq(pd_data == pd_data, gdf_data == gdf_data)
-    assert_eq(pd_data < pd_data, gdf_data < gdf_data)
-    assert_eq(pd_data > pd_data, gdf_data > gdf_data)
-
-
-@pytest.mark.skipif(
-    PANDAS_VERSION < PANDAS_CURRENT_SUPPORTED_VERSION,
-    reason="Fails in older versions of pandas.",
-)
-def test_datetime_can_cast_safely():
-    sr = cudf.Series(
-        ["1679-01-01", "2000-01-31", "2261-01-01"], dtype="datetime64[ms]"
-    )
-    assert sr._column.can_cast_safely(np.dtype("datetime64[ns]"))
-
-    sr = cudf.Series(
-        ["1677-01-01", "2000-01-31", "2263-01-01"], dtype="datetime64[ms]"
-    )
-
-    assert sr._column.can_cast_safely(np.dtype("datetime64[ns]")) is False
-
-
-@pytest.mark.parametrize(
-    "data",
-    [
-        [1, 2, 3, 4, 10, 100, 20000],
-        [None] * 7,
-        [10, 20, 30, None, 100, 200, None],
-        [3223.234, 342.2332, 23423.23, 3343.23324, 23432.2323, 242.23, 233],
-    ],
-)
-@pytest.mark.parametrize(
-    "other",
-    [
-        [1, 2, 3, 4, 10, 100, 20000],
-        [None] * 7,
-        [10, 20, 30, None, 100, 200, None],
-        [3223.234, 342.2332, 23423.23, 3343.23324, 23432.2323, 242.23, 233],
-        datetime.datetime(1993, 6, 22, 13, 30),
-        datetime.datetime(2005, 1, 22, 10, 00),
-        np.datetime64("2005-02"),
-        np.datetime64("2005-02-25"),
-        np.datetime64("2005-02-25T03:30"),
-        np.datetime64("nat"),
-        # TODO: https://github.com/pandas-dev/pandas/issues/52295
-    ],
-)
-@pytest.mark.parametrize("data_dtype", DATETIME_TYPES)
-@pytest.mark.parametrize("other_dtype", DATETIME_TYPES)
-def test_datetime_subtract(data, other, data_dtype, other_dtype):
-    gsr = cudf.Series(data, dtype=data_dtype)
-    psr = gsr.to_pandas()
-
-    if isinstance(other, np.datetime64):
-        gsr_other = other
-        psr_other = other
-    else:
-        gsr_other = cudf.Series(other, dtype=other_dtype)
-        psr_other = gsr_other.to_pandas()
-
-    expected = psr - psr_other
-    actual = gsr - gsr_other
-
-    assert_eq(expected, actual)
-
-    expected = psr_other - psr
-    actual = gsr_other - gsr
-
-    assert_eq(expected, actual)
-
-
-@pytest.mark.parametrize(
-    "data",
-    [
-        [1000000, 200000, 3000000],
-        [1000000, 200000, None],
-        [],
-        [None],
-        [None, None, None, None, None],
-        [12, 12, 22, 343, 4353534, 435342],
-        np.array([10, 20, 30, None, 100]),
-        cp.asarray([10, 20, 30, 100]),
-        [1000000, 200000, 3000000],
-        [1000000, 200000, None],
-        [1],
-        [12, 11, 232, 223432411, 2343241, 234324, 23234],
-        [12, 11, 2.32, 2234.32411, 2343.241, 23432.4, 23234],
-        [1.321, 1132.324, 23223231.11, 233.41, 0.2434, 332, 323],
-        [12, 11, 2.32, 2234.32411, 2343.241, 23432.4, 23234],
-    ],
-)
-@pytest.mark.parametrize(
-    "other_scalars",
-    [
-        datetime.timedelta(days=768),
-        datetime.timedelta(seconds=768),
-        datetime.timedelta(microseconds=7),
-        datetime.timedelta(minutes=447),
-        datetime.timedelta(hours=447),
-        datetime.timedelta(weeks=734),
-        np.timedelta64(4, "s"),
-        np.timedelta64(456, "D"),
-        np.timedelta64(46, "h"),
-        np.timedelta64("nat"),
-        np.timedelta64(1, "s"),
-        np.timedelta64(1, "ms"),
-        np.timedelta64(1, "us"),
-        np.timedelta64(1, "ns"),
-    ],
-)
-@pytest.mark.parametrize("dtype", DATETIME_TYPES)
-@pytest.mark.parametrize(
-    "op",
-    ["add", "sub"],
-)
-def test_datetime_series_ops_with_scalars(data, other_scalars, dtype, op):
-    gsr = cudf.Series(data=data, dtype=dtype)
-    psr = gsr.to_pandas()
-
-    if op == "add":
-        expected = psr + other_scalars
-        actual = gsr + other_scalars
-    elif op == "sub":
-        expected = psr - other_scalars
-        actual = gsr - other_scalars
-
-    assert_eq(expected, actual)
-
-    if op == "add":
-        expected = other_scalars + psr
-        actual = other_scalars + gsr
-
-        assert_eq(expected, actual)
-
-    elif op == "sub":
-        assert_exceptions_equal(
-            lfunc=operator.sub,
-            rfunc=operator.sub,
-            lfunc_args_and_kwargs=([other_scalars, psr],),
-            rfunc_args_and_kwargs=([other_scalars, gsr],),
-        )
-
-
-@pytest.mark.parametrize("data", ["20110101", "20120101", "20130101"])
-@pytest.mark.parametrize("other_scalars", ["20110101", "20120101", "20130101"])
-@pytest.mark.parametrize(
-    "dtype",
-    ["datetime64[ns]", "datetime64[us]", "datetime64[ms]", "datetime64[s]"],
-)
-def test_datetime_series_cmpops_with_scalars(data, other_scalars, dtype, op):
-    gsr = cudf.Series(data=data, dtype=dtype)
-    psr = gsr.to_pandas()
-
-    expect = op(psr, other_scalars)
-    got = op(gsr, other_scalars)
-
-    assert_eq(expect, got)
-
-
-@pytest.mark.parametrize(
-    "data",
-    [
-        [1000000, 200000, 3000000],
-        [1000000, 200000, None],
-        [],
-        [None],
-        [None, None, None, None, None],
-        [12, 12, 22, 343, 4353534, 435342],
-        np.array([10, 20, 30, None, 100]),
-        cp.asarray([10, 20, 30, 100]),
-        [1000000, 200000, 3000000],
-        [1000000, 200000, None],
-        [1],
-        [12, 11, 232, 223432411, 2343241, 234324, 23234],
-        [12, 11, 2.32, 2234.32411, 2343.241, 23432.4, 23234],
-        [1.321, 1132.324, 23223231.11, 233.41, 0.2434, 332, 323],
-        [12, 11, 2.32, 2234.32411, 2343.241, 23432.4, 23234],
-    ],
-)
-@pytest.mark.parametrize(
-    "scalar",
-    [
-        datetime.timedelta(days=768),
-        datetime.timedelta(seconds=768),
-        datetime.timedelta(microseconds=7),
-        np.timedelta64("nat"),
-        np.timedelta64(1, "s"),
-        np.timedelta64(1, "ms"),
-        np.timedelta64(1, "us"),
-        np.timedelta64(1, "ns"),
-    ],
-)
-@pytest.mark.parametrize("dtype", DATETIME_TYPES)
-@pytest.mark.parametrize("op", [np.add, np.subtract])
-def test_datetime_series_ops_with_scalars_misc(data, scalar, dtype, op):
-    gsr = cudf.Series(data=data, dtype=dtype)
-    psr = gsr.to_pandas()
-
-    expect = op(psr, scalar)
-    got = op(gsr, scalar)
-
-    assert_eq(expect, got)
-
-
-def test_datetime_invalid_ops():
-    sr = cudf.Series([1, 2, 3], dtype="datetime64[ns]")
-    psr = sr.to_pandas()
-
-    assert_exceptions_equal(
-        lfunc=operator.add,
-        rfunc=operator.add,
-        lfunc_args_and_kwargs=([psr, pd.Timestamp(1513393355.5, unit="s")],),
-        rfunc_args_and_kwargs=([sr, pd.Timestamp(1513393355.5, unit="s")],),
-    )
-
-    assert_exceptions_equal(
-        lfunc=operator.truediv,
-        rfunc=operator.truediv,
-        lfunc_args_and_kwargs=([psr, pd.Timestamp(1513393355.5, unit="s")],),
-        rfunc_args_and_kwargs=([sr, pd.Timestamp(1513393355.5, unit="s")],),
-    )
-
-    assert_exceptions_equal(
-        lfunc=operator.add,
-        rfunc=operator.add,
-        lfunc_args_and_kwargs=([psr, psr],),
-        rfunc_args_and_kwargs=([sr, sr],),
-    )
-
-    assert_exceptions_equal(
-        lfunc=operator.floordiv,
-        rfunc=operator.floordiv,
-        lfunc_args_and_kwargs=([psr, psr],),
-        rfunc_args_and_kwargs=([sr, sr],),
-    )
-
-    assert_exceptions_equal(
-        lfunc=operator.floordiv,
-        rfunc=operator.floordiv,
-        lfunc_args_and_kwargs=([psr, pd.Timestamp(1513393355.5, unit="s")],),
-        rfunc_args_and_kwargs=([sr, pd.Timestamp(1513393355.5, unit="s")],),
-    )
-
-    assert_exceptions_equal(
-        lfunc=operator.add,
-        rfunc=operator.add,
-        lfunc_args_and_kwargs=([psr, 1],),
-        rfunc_args_and_kwargs=([sr, 1],),
-    )
-
-    assert_exceptions_equal(
-        lfunc=operator.truediv,
-        rfunc=operator.truediv,
-        lfunc_args_and_kwargs=([psr, "a"],),
-        rfunc_args_and_kwargs=([sr, "a"],),
-    )
-
-    assert_exceptions_equal(
-        lfunc=operator.mul,
-        rfunc=operator.mul,
-        lfunc_args_and_kwargs=([psr, 1],),
-        rfunc_args_and_kwargs=([sr, 1],),
-    )
-
-
-def test_datetime_binop_tz_timestamp(op):
-    s = cudf.Series([1, 2, 3], dtype="datetime64[ns]")
-    pd_tz_timestamp = pd.Timestamp("1970-01-01 00:00:00.000000001", tz="utc")
-    with pytest.raises(NotImplementedError):
-        op(s, pd_tz_timestamp)
-
-    date_scalar = datetime.datetime.now(datetime.timezone.utc)
-    with pytest.raises(NotImplementedError):
-        op(s, date_scalar)
-
-
-def test_datetime_series_cmpops_pandas_compatibility(op):
-    data1 = ["20110101", "20120101", None, "20140101", None]
-    data2 = ["20110101", "20120101", "20130101", None, None]
-    gsr1 = cudf.Series(data=data1, dtype="datetime64[ns]")
-    psr1 = gsr1.to_pandas()
-
-    gsr2 = cudf.Series(data=data2, dtype="datetime64[ns]")
-    psr2 = gsr2.to_pandas()
-
-    expect = op(psr1, psr2)
-    with cudf.option_context("mode.pandas_compatible", True):
-        got = op(gsr1, gsr2)
-
-    assert_eq(expect, got)
-
-
-@pytest.mark.parametrize(
-    "method, kwargs",
-    [["mean", {}], ["std", {}], ["std", {"ddof": 0}]],
-)
-def test_dti_reduction(method, kwargs):
-    pd_dti = pd.DatetimeIndex(["2020-01-01", "2020-12-31"], name="foo")
-    cudf_dti = cudf.from_pandas(pd_dti)
-
-    result = getattr(cudf_dti, method)(**kwargs)
-    expected = getattr(pd_dti, method)(**kwargs)
-    assert result == expected
